split symm_basic.cpp (1/4): extract atom_ordering_new/test_atom_ordering to symm_basic_order.cpp
